+++ vue.config.js
@@ -7,9 +7,9 @@
 const path = require('path');
 const resolve = (dir) => path.join(__dirname, dir);
 module.exports = {
-  publicPath: '/admin', // 部署应用时的根路径(默认'/'),也可用相对路径(当使用相对路径时，router中也需要配置base为对应的路径)
+  publicPath: './', // 部署应用时的根路径(默认'/'),也可用相对路径'./'(当使用相对路径时，router中也需要配置base为对应的路径)
   outputDir: 'dist', // 运行时生成的生产环境构建文件的目录(默认''dist''，构建之前会被清除)
-  assetsDir: 'public', // 放置生成的静态资源(s、css、img、fonts)的(相对于 outputDir 的)目录(默认'')
+  assetsDir: './public', // 放置生成的静态资源(s、css、img、fonts)的(相对于 outputDir 的)目录(默认'')
   indexPath: 'index.html', // 指定生成的 index.html 的输出路径(相对于 outputDir)也可以是一个绝对路径。
   pages: {
     // pages 里配置的路径和文件名在你的文档目录必须存在 否则启动服务会报错